fix(assisted): added size prop
