+++ package.json
@@ -37,7 +37,6 @@
   "dependencies": {
     "@inubekit/button": "^1.0.0",
     "@inubekit/grid": "^1.0.0",
-    "@inubekit/hooks": "^1.1.0",
     "@inubekit/icon": "^1.1.1",
     "@inubekit/stack": "^1.1.0",
     "@inubekit/text": "^1.3.1"
