+++ src/Assisted/stories/Assisted.stories.tsx
@@ -53,6 +53,7 @@ const Default = (args: IAssisted) => <AssistedController {...args} />;
 
 Default.args = {
   steps: stepsMock,
+  size: "small",
   currentStepId: 3,
   titleButtonText: {
     before: "Anterior",
